Split iOS review queue package
