+++ apps/ios/Flashcards/Flashcards/Review/Queue/Runtime/ReviewQueueRuntime+CountLoading.swift
@@ -1,0 +1,77 @@
+import Foundation
+
+struct ReviewCountsLoadRequest {
+    let databaseURL: URL
+    let workspaceId: String
+    let reviewQueryDefinition: ReviewQueryDefinition
+    let now: Date
+    let requestId: String
+    let sourceVersion: Int
+}
+
+extension ReviewQueueRuntime {
+    mutating func startReviewCountsLoad(request: ReviewCountsLoadRequest) {
+        self.cancelActiveReviewCountsLoad()
+        self.state.activeReviewCountsRequestId = request.requestId
+    }
+
+    mutating func setActiveReviewCountsTask(task: Task<Void, Never>, requestId: String) {
+        self.state.activeReviewCountsTask = task
+        self.state.activeReviewCountsRequestId = requestId
+    }
+
+    mutating func applyReviewCountsLoadSuccess(
+        publishedState: ReviewQueuePublishedState,
+        reviewCounts: ReviewCounts,
+        requestId: String,
+        sourceVersion: Int
+    ) -> ReviewQueuePublishedState? {
+        guard self.shouldApplyReviewCountsResult(requestId: requestId, sourceVersion: sourceVersion) else {
+            return nil
+        }
+
+        self.clearActiveReviewCountsLoad(requestId: requestId)
+        return ReviewQueuePublishedState(
+            selectedReviewFilter: publishedState.selectedReviewFilter,
+            reviewQueue: publishedState.reviewQueue,
+            presentedReviewCard: self.resolvePresentedReviewCard(
+                reviewQueue: publishedState.reviewQueue,
+                pendingReviewCardIds: publishedState.pendingReviewCardIds,
+                preferredPresentedReviewCard: publishedState.presentedReviewCard
+            ),
+            reviewCounts: reviewCounts,
+            isReviewHeadLoading: publishedState.isReviewHeadLoading,
+            isReviewCountsLoading: false,
+            isReviewQueueChunkLoading: publishedState.isReviewQueueChunkLoading,
+            pendingReviewCardIds: publishedState.pendingReviewCardIds,
+            reviewSubmissionFailure: publishedState.reviewSubmissionFailure
+        )
+    }
+
+    mutating func applyReviewCountsLoadFailure(
+        publishedState: ReviewQueuePublishedState,
+        requestId: String,
+        sourceVersion: Int
+    ) -> ReviewQueuePublishedState? {
+        guard self.shouldApplyReviewCountsResult(requestId: requestId, sourceVersion: sourceVersion) else {
+            return nil
+        }
+
+        self.clearActiveReviewCountsLoad(requestId: requestId)
+        return ReviewQueuePublishedState(
+            selectedReviewFilter: publishedState.selectedReviewFilter,
+            reviewQueue: publishedState.reviewQueue,
+            presentedReviewCard: self.resolvePresentedReviewCard(
+                reviewQueue: publishedState.reviewQueue,
+                pendingReviewCardIds: publishedState.pendingReviewCardIds,
+                preferredPresentedReviewCard: publishedState.presentedReviewCard
+            ),
+            reviewCounts: publishedState.reviewCounts,
+            isReviewHeadLoading: publishedState.isReviewHeadLoading,
+            isReviewCountsLoading: false,
+            isReviewQueueChunkLoading: publishedState.isReviewQueueChunkLoading,
+            pendingReviewCardIds: publishedState.pendingReviewCardIds,
+            reviewSubmissionFailure: publishedState.reviewSubmissionFailure
+        )
+    }
+}
