+++ apps/web/src/appData/domain/index.ts
@@ -344,7 +344,7 @@ function getReviewOrderCreatedTimestamp(card: Card): number {
 
 /**
  * Keep review queue ordering aligned with:
- * - apps/ios/Flashcards/Flashcards/Review/Queue/ReviewQuerySupport.swift::compareCardsForReviewOrder
+ * - apps/ios/Flashcards/Flashcards/Review/Queue/Query/ReviewQuerySupport.swift::compareCardsForReviewOrder
  * - apps/ios/Flashcards/Flashcards/Database/CardStore/CardStore+ReadSQL.swift review queue ORDER BY
  * - apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/review/ReviewSupport.kt::sortCardsForReviewQueue
  * Ordering contract: recently reviewed due cards in the inclusive 1-hour fsrsLastReviewedAt
